Factor out app middleware
