+++ routes/params.js
@@ -1,0 +1,6 @@
+module.exports = function (app) {
+  app.param('format', function (req, res, next){
+    req.params.format = req.params.format.toLowerCase();
+    next();
+  });
+};
